Refactor averaging module into mixin files
